+++ cli/aux/pwd.c
@@ -8,15 +8,15 @@ int tec_cli_pwd_set(tec_arg_t *args)
     FILE *fp;
     char *taskdir = teccfg.base.task;
     const char *fmt = "%s/%s/%s/%s\n";
-    const char *fmtdebug = "pwd: env='%s', desk='%s', taskid='%s'";
+    const char *fmtdebug = "pwd: env='%s', desk='%s', task='%s'";
 
     args->env = args->env == NULL ? "" : args->env;
     args->desk = args->desk == NULL ? "" : args->desk;
-    args->taskid = args->taskid == NULL ? "" : args->taskid;
+    args->task = args->task == NULL ? "" : args->task;
 
     if ((fp = fopen(PWDFILE, "w"))) {
-        dlog(1, fmtdebug, args->env, args->desk, args->taskid);
-        fprintf(fp, fmt, taskdir, args->env, args->desk, args->taskid);
+        dlog(1, fmtdebug, args->env, args->desk, args->task);
+        fprintf(fp, fmt, taskdir, args->env, args->desk, args->task);
         return fclose(fp);
     }
     return 1;