+++ cli/aux/toggle.h
@@ -22,6 +22,6 @@ int toggle_task_unset_prev(char *base, tec_arg_t * args);
 /* Update toggles after rename/move operations */
 int toggle_task_update(char *base, tec_arg_t * args,
                        const char *old_id, const char *new_id);
-int toggle_task_clear(char *base, tec_arg_t * args, const char *taskid);
+int toggle_task_clear(char *base, tec_arg_t * args, const char *task);
 
 #endif